Moved reftypes.py from core to certify since it is most closely related to certify.   Fixed missing self.locator attribute in reftypes.py
